+++ DataAccess/Concrete/EntityFramework/EfUserAddressDal.cs
@@ -26,7 +26,7 @@ namespace DataAccess.Concrete.EntityFramework
                                  Id = adress.Id,
                                  UserId = adress.UserId,
                                  AddressName = adress.AddressName,
-                                 CityId = adress.CityId,
+                                 CityId = (byte)adress.CityId,
                                  AddressText = adress.AddressText,
                                  CityName = city.CityName
                              };
